# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/com/youcode/visionarycrofting/repository/CommandItemRepository.java
@@ -3,5 +3,5 @@ package com.youcode.visionarycrofting.repository;
 import com.youcode.visionarycrofting.entity.CommandItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommandItemRepository extends JpaRepository< CommandItem, Integer > {
+public interface CommandItemRepository extends JpaRepository< CommandItem, Long > {
 }
